requirements.txt
requests
pandas
numpy
eel
pyinstaller
python-dotenv
